Fix for SummernoteInplaceWidget in widgets.py
SummernoteInplaceWidget does not work if widget context contains __proxy__ values obtained via ugettext_lazy. This quick hack checks the type of the context value and if it's of type __proxy__ forces unicode evaluation so that it can be properly JSON serialized. Without this check the following error this thrown and the Widget will not render.

`TypeError: Object of type '__proxy__' is not JSON serializable`